% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R, R/zzz.R
\name{LibhvCompileVersion}
\alias{LibhvCompileVersion}
\title{Get the libhv Compile Version}
\usage{
LibhvCompileVersion()

LibhvCompileVersion()
}
\value{
A character string with the compile-time version of libhv

A character string with the compile-time version of libhv
}
\description{
Returns the compile-time version string of the libhv library

Returns the compile-time version string of the libhv library
}
