% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Info.R
\name{LibhvInfo}
\alias{LibhvInfo}
\title{Rlibhv Package Information}
\usage{
LibhvInfo()
}
\value{
A list containing version information for libhv and its dependencies
}
\description{
Returns information about the Rlibhv package including the libhv version
and other dependencies.
}
\examples{
LibhvInfo()
}
