% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R, R/zzz.R
\name{LibhvVersion}
\alias{LibhvVersion}
\title{Get the libhv Version}
\usage{
LibhvVersion()

LibhvVersion()
}
\value{
A character string with the version of libhv

A character string with the version of libhv
}
\description{
Returns the version string of the libhv library

Returns the version string of the libhv library
}
