% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R, R/zzz.R
\name{LibhvVersionNumber}
\alias{LibhvVersionNumber}
\title{Get the libhv Version Number}
\usage{
LibhvVersionNumber()

LibhvVersionNumber()
}
\value{
An integer representing the version number of libhv

An integer representing the version number of libhv
}
\description{
Returns the version number of the libhv library as an integer

Returns the version number of the libhv library as an integer
}
