% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageConfig.R
\name{packageCflags}
\alias{packageCflags}
\title{Package CFLAGS for compilation}
\usage{
packageCflags()
}
\value{
A string containing the CFLAGS for compiling with libhv and its dependencies
}
\description{
Package CFLAGS for compilation
}
\examples{
\dontrun{
# Get the CFLAGS for compiling with libhv
packageCflags()
}
}
