% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageConfig.R
\name{packageLibs}
\alias{packageLibs}
\title{Package libraries for linking}
\usage{
packageLibs()
}
\value{
A string containing the libraries to link against
}
\description{
Package libraries for linking
}
\examples{
\dontrun{
# Get the libraries for linking with libhv
packageLibs()
}
}
